backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-PyMongo==2.3.0
flasgger==0.9.7.1
pymongo==4.6.0
marshmallow==3.21.1
requests==2.31.0
openai==1.12.0
sudachipy==0.6.10
sudachidict-core==20230927
python-dotenv==1.0.0
pydantic==2.12.3
pydantic-settings==2.1.0
injector==0.21.0
gunicorn==21.2.0
pytest==7.4.3
ruff==0.6.8
pre-commit==3.8.0
PyJWT==2.8.0
